Cleanup responses a bit more
