--- datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v0/models/GetRecipeIngredientResponseV0.kt
@@ -1,14 +1,0 @@
-package gq.kirmanak.mealient.datasource.v0.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GetRecipeIngredientResponseV0(
-    @SerialName("title") val title: String = "",
-    @SerialName("note") val note: String = "",
-    @SerialName("unit") val unit: String = "",
-    @SerialName("food") val food: String = "",
-    @SerialName("disableAmount") val disableAmount: Boolean,
-    @SerialName("quantity") val quantity: Double,
-)
