--- datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v0/models/GetRecipeInstructionResponseV0.kt
@@ -1,10 +1,0 @@
-package gq.kirmanak.mealient.datasource.v0.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GetRecipeInstructionResponseV0(
-    @SerialName("title") val title: String = "",
-    @SerialName("text") val text: String,
-)
