+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v0/models/GetTokenResponseV0.kt
@@ -4,4 +4,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetTokenResponseV0(@SerialName("access_token") val accessToken: String)
+data class GetTokenResponseV0(
+    @SerialName("access_token") val accessToken: String,
+)
