+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v0/models/VersionResponseV0.kt
@@ -5,7 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class VersionResponseV0(
-    @SerialName("production") val production: Boolean,
     @SerialName("version") val version: String,
-    @SerialName("demoStatus") val demoStatus: Boolean,
 )
