+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v1/models/ErrorDetailV1.kt
@@ -4,4 +4,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ErrorDetailV1(@SerialName("detail") val detail: String? = null)
+data class ErrorDetailV1(
+    @SerialName("detail") val detail: String? = null,
+)
