+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v1/models/GetTokenResponseV1.kt
@@ -6,5 +6,4 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GetTokenResponseV1(
     @SerialName("access_token") val accessToken: String,
-    @SerialName("token_type") val tokenType: String,
 )
