+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/v1/models/VersionResponseV1.kt
@@ -5,8 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class VersionResponseV1(
-    @SerialName("production") val production: Boolean,
     @SerialName("version") val version: String,
-    @SerialName("demoStatus") val demoStatus: Boolean,
-    @SerialName("allowSignup") val allowSignup: Boolean,
 )
